// src/domain/course/entity/DestinationPreference.enum.ts
/**
 * DestinationPreference Enum
 * @description
 * - 여행지 선호 열거형
 * - 로드맵 생성 설문: 관광지 위주 vs 로컬 위주
 */
export enum DestinationPreference {
  TOURIST_SPOTS = 'TOURIST_SPOTS', // 관광지 위주
  LOCAL_EXPERIENCE = 'LOCAL_EXPERIENCE', // 로컬 위주
}