// src/domain/course/entity/TravelTheme.enum.ts
/**
 * TravelTheme Enum
 * @description
 * - 여행 테마 열거형
 * - 로드맵 생성 설문: 여행 컨셉
 */
export enum TravelTheme {
  UNIQUE_TRIP = 'UNIQUE_TRIP', // 이색여행
  HONEYMOON = 'HONEYMOON', // 신혼여행
  FAMILY_TRIP = 'FAMILY_TRIP', // 가족여행
  HEALING = 'HEALING', // 힐링
  SIGHTSEEING = 'SIGHTSEEING', // 관광
  FOOD_TOUR = 'FOOD_TOUR', // 먹방
  NATURE = 'NATURE', // 자연
  SHOPPING = 'SHOPPING', // 쇼핑
  CULTURE_ART = 'CULTURE_ART', // 문화·예술
  ACTIVITY = 'ACTIVITY', // 액티비티
  CITY_TRIP = 'CITY_TRIP', // 도시여행
  PHOTO_SPOTS = 'PHOTO_SPOTS', // 사진·인생샷
}