// src/domain/course/entity/Companion.enum.ts
/**
 * Companion Enum
 * @description
 * - 동행자 유형 열거형
 * - 로드맵 생성 설문: 동행자
 */
export enum Companion {
  FAMILY = 'FAMILY', // 가족
  FRIENDS = 'FRIENDS', // 친구
  COUPLE = 'COUPLE', // 연인
  CHILDREN = 'CHILDREN', // 아이
  PARENTS = 'PARENTS', // 부모님
  COLLEAGUES = 'COLLEAGUES', // 직장 동료
  SOLO = 'SOLO', // 혼자
}