// src/domain/place/entity/PlaceCategory.enum.ts
export enum PlaceCategory {
  FOOD = 'FOOD',
  CAFE = 'CAFE',
  ATTRACTION = 'ATTRACTION',
  CULTURE = 'CULTURE',
  NATURE = 'NATURE',
  SHOPPING = 'SHOPPING',
  LODGING = 'LODGING',
  TRANSPORT = 'TRANSPORT',
  ENTERTAINMENT = 'ENTERTAINMENT',
  WELLNESS = 'WELLNESS',
  SERVICE = 'SERVICE',
  OTHER = 'OTHER',
}